Bug 773296 - Part 3: Allow more than 27 style structs. r=dbaron
This bumps up nsStyleContext::mBits to a uint64_t so that it can fit
another style struct.  If we're going to need to keep at least 27 style
structs, it might be better to split mBits up into two uint32_ts: one
for the flags and one for the style struct bits.

UltraBlame original commit: 7069d52b1e601f6bc55b4609e055233b7f5f734d

+++ layout/style/nsStyleContext.h
@@ -211,7 +211,7 @@ public:
   #undef STYLE_STRUCT_INHERITED
 
   nsRuleNode* RuleNode() { return mRuleNode; }
-  void AddStyleBit(const uint32_t& aBit) { mBits |= aBit; }
+  void AddStyleBit(const uint64_t& aBit) { mBits |= aBit; }
 
   
 
@@ -430,7 +430,7 @@ protected:
   
   nsResetStyleData*       mCachedResetData; 
   nsInheritedStyleData    mCachedInheritedData; 
-  uint32_t                mBits; 
+  uint64_t                mBits; 
                                  
   uint32_t                mRefCnt;
 };
